Updated version Phasetech2DIR
User works with "Main screen"
